% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simudata.R
\name{DGP_mcp}
\alias{DGP_mcp}
\title{Simulate a piecewise stationary  process}
\usage{
DGP_mcp(TT, m, sparsity, q, type)
}
\arguments{
\item{TT}{data length.}

\item{m}{data dimension.}

\item{sparsity}{number of series activated by change. Cp series are evenly distributed across the \code{m} dimensions.}

\item{q}{number of change points, evenly distributed across the \code{TT} length.}

\item{type}{VMA or VAR.}
}
\value{
the data \code{y}.
}
\description{
Simulate piecewise stationary VAR or VMA processes with partial series change and multiple change points. They are used in the simulation studies of Zhang and Chan (2025).
}
\examples{
set.seed(1)
y <- DGP_mcp(TT = 6000, m = 80, sparsity = 8, q = 4, type = 'vma')
}
\references{
Zhang and Chan. (2025). Spectral change point estimation for high dimensional time series by sparse tensor decomposition.
}
