% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{find.b}
\alias{find.b}
\title{Function to find the maximum point of a positive vector, considering only points around which an interval of length  \eqn{2\delta} are all positive}
\usage{
find.b(stat, delta)
}
\arguments{
\item{stat}{the positive vector}

\item{delta}{let the picked location within a interval with positive projected cusum, of length \eqn{2\delta}}
}
\value{
a list containing the maximum point and the maximum value
}
\description{
Function to find the maximum point of a positive vector, considering only points around which an interval of length  \eqn{2\delta} are all positive
}
\keyword{internal}
