% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tensordecomp.R
\name{myalternative}
\alias{myalternative}
\title{Function to get sparse leading term of a partial symmetric (at mode 1 and 2) partial constant (at mode 3) tensor via power method}
\usage{
myalternative(T, a0, sparsity = NULL, niter = 10)
}
\arguments{
\item{T}{the three order tensor}

\item{a0}{the initial vector at mode 1}

\item{sparsity}{sparsity value at mode 1}

\item{niter}{maximum number of iterations}
}
\description{
Function to get sparse leading term of a partial symmetric (at mode 1 and 2) partial constant (at mode 3) tensor via power method
}
\keyword{internal}
