% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tensordecomp.R
\name{myinitial_agg}
\alias{myinitial_agg}
\title{Function for initial vectors based on aggregation}
\usage{
myinitial_agg(T, sparsity = NULL)
}
\arguments{
\item{T}{the three order tensor}

\item{sparsity}{sparsity value at mode 1}

\item{niter}{maximum number of iterations}
}
\description{
Function for initial vectors based on aggregation
}
\keyword{internal}
