% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tensordecomp.R
\name{mystensor}
\alias{mystensor}
\title{Main function for sparse tensor decomposi tion}
\usage{
mystensor(T, sparsity, niter = 10, initial = "aggregation")
}
\arguments{
\item{T}{the three order tensor}

\item{sparsity}{sparsity value at mode 1}

\item{niter}{maximum number of iterations}

\item{initial}{the initialization method, can be either aggregation or unfolding.}
}
\description{
Main function for sparse tensor decomposi tion
}
\keyword{internal}
