% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tensordecomp.R
\name{mytpower}
\alias{mytpower}
\title{Function of getting the sparse leading eigen vector}
\usage{
mytpower(
  X,
  sparsity,
  v1.init,
  niter = 10,
  lambda.diff.threshold = 1e-06,
  trace = FALSE
)
}
\arguments{
\item{X}{symmetric matrix}

\item{sparsity}{sparsity value}

\item{v1.init}{the initial vector}
}
\description{
Function of getting the sparse leading eigen vector
}
\keyword{internal}
