% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{NormalTrans}
\alias{NormalTrans}
\title{Function to apply the normal density transformation.
This function is helpful to relieve the outlier/heavy tail effect.}
\usage{
NormalTrans(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a transformed vector
}
\description{
Function to apply the normal density transformation.
This function is helpful to relieve the outlier/heavy tail effect.
}
\keyword{internal}
