% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{sparsum}
\alias{sparsum}
\title{Function of sparsified sum over frequencies for multivariate time series
@param output the output by \code{SpecProj}, if not given, will be computed with \code{bs.part}
@param bs.part the four order spectral tensor, of dim by dim by freq (or band) by block
@param m the dimension, since univariate and multivariate cases are different that the former does not need projection}
\usage{
sparsum(
  output = NULL,
  bs.part = NULL,
  m,
  sparsity = NULL,
  tau,
  trim = NULL,
  initial = "aggregation"
)
}
\arguments{
\item{sparsity}{the sparsity value}

\item{tau}{the threshold}

\item{trim}{the boundary removal \eqn{\nu}}

\item{initial}{the initialization method, can be either aggregation or unfolding.}
}
\value{
a list including the maximum cusum, the maximum location, and informations of the series and frequency.
}
\description{
Function of sparsified sum over frequencies for multivariate time series
@param output the output by \code{SpecProj}, if not given, will be computed with \code{bs.part}
@param bs.part the four order spectral tensor, of dim by dim by freq (or band) by block
@param m the dimension, since univariate and multivariate cases are different that the former does not need projection
}
\keyword{internal}
