% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{SpecBlock}
\alias{SpecBlock}
\title{Spectral matrix estimate in blocks}
\usage{
SpecBlock(
  y,
  L,
  freq = seq(0, pi, length.out = 19)[-1],
  R = floor(L^(1/3)),
  band = rep(1, length(freq)),
  scale = TRUE,
  normaltrans = TRUE,
  coherency = FALSE
)
}
\arguments{
\item{y}{A \code{T} by \code{m} matrix for a multivariate time series of dimension \code{m} and length \code{T} .}

\item{L}{The block size, a scalar \verb{0<L<T}. Could be set subject to user preference. At least 50 for a good estimate of spectrum within block.}

\item{freq}{The frequencies on which the spectral density. will be estimated, can be specified due to expert knowledge or user interest. The default values are 18 values evenly distributed between 0 and pi.}

\item{R}{(optional) The bandwidth, a scalar \verb{0<R<L} with default \code{R= floor(L^{1/3})}.}

\item{band}{(optional) Use it when users want to assess frequency bands instead of individual frequencies. An integer vector indicating the number in each band. Specifically, the vector is of length \eqn{a} and sum to \eqn{b}, where \eqn{a} is the number of groups, and \eqn{b}  equals to length of freq. Default value means no band.}

\item{scale}{(optional, default \code{TRUE}) Logical indicating whether marginal scale will be implemented before estimation.}

\item{normaltrans}{(optional, default \code{TRUE}) Logical indicating whether a marginal nonparametric normal transformation will be implemented before estimation.}

\item{coherency}{(optional, default \code{FALSE})  Use it if users want to assess the coherency matrix, instead of spectral matrix. If \code{TRUE}, the spectral matrix will be normalized such that it has unit diagonal.}
}
\value{
Returns the spectral estimate, which is a four-dimensional array, of dim by dim by freq (or band) by block
}
\description{
Lag window estimate of spectral matrix in non-overlapping blocks.
}
\author{
Xinyu Zhang xinyu-zhang@uiowa.edu
}
