% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{SpecCp}
\alias{SpecCp}
\title{Change point estimation for multivariate time series from frequency domain}
\usage{
SpecCp(
  y,
  L,
  freq = seq(0, pi, length.out = 19)[-1],
  R = max(1, floor(L^(1/3))),
  band = rep(1, length(freq)),
  sparsity = NULL,
  refine = TRUE,
  do.parallel = 0,
  initial = "aggregation",
  mean.output = FALSE,
  tau = NULL,
  M = NULL,
  Nb = 100,
  scale = TRUE,
  normaltrans = TRUE,
  coherency = FALSE,
  single = FALSE,
  message = TRUE
)
}
\arguments{
\item{y}{A \code{T} by \code{m} matrix for a multivariate time series of dimension \code{m} and length \code{T} .}

\item{L}{The block size, a scalar \verb{0<L<T}. Could be set subject to user preference. At least 50 for a good estimate of spectrum within block.}

\item{freq}{The frequencies on which the spectral density. will be estimated, can be specified due to expert knowledge or user interest. The default values are 18 values evenly distributed between 0 and pi.}

\item{R}{(optional) The bandwidth, a scalar \verb{0<R<L} with default \code{R= floor(L^{1/3})}.}

\item{band}{(optional) Use it when users want to assess frequency bands instead of individual frequencies. An integer vector indicating the number in each band. Specifically, the vector is of length \eqn{a} and sum to \eqn{b}, where \eqn{a} is the number of groups, and \eqn{b}  equals to length of freq. Default value means no band.}

\item{sparsity}{(optional) The sparsity value. Default values are estimated from the data based on the data using one-by-one approach, so users do not need to specify unless they want to override the defaults.}

\item{refine}{(optional, default \code{TRUE}) Logical indicating whether a refinement step will be indicated by scanning the localized blocks around the change point detected for multivariate time series}

\item{do.parallel}{(optional, default 0) Number of copies of R running in parallel}

\item{initial}{the initialization method, can be either aggregation or unfolding.}

\item{mean.output}{(optional, default \code{FALSE}) Logical indicating whether the piecewise mean spectral will be output}

\item{tau}{(optional) The threshold. Default values are estimated from the data, so users do not need to specify unless they want to override the defaults.}

\item{M}{(optional) The number of random intervals in the wild binary segmentation.}

\item{Nb}{(optional) The number of bootstraps for obtaining the threshold value.}

\item{scale}{(optional, default \code{TRUE}) Logical indicating whether marginal scale will be implemented before estimation.}

\item{normaltrans}{(optional, default \code{TRUE}) Logical indicating whether a marginal nonparametric normal transformation will be implemented before estimation.}

\item{coherency}{(optional, default \code{FALSE})  Use it if users want to assess the coherency matrix, instead of spectral matrix. If \code{TRUE}, the spectral matrix will be normalized such that it has unit diagonal.}

\item{single}{(optional, default \code{FALSE}) If \code{TRUE}, only a single change point will be estimated, and binary segmentation will not be implemented. Should be omitted and only be used for a fast conclusion of breaks existence.}

\item{message}{(optional, default \code{TRUE}) If \code{TRUE}, print out the computation process.}
}
\value{
A list with the following elements
\describe{
\item{cp}{A matrix with \code{k} rows, where \code{k} is the number of detected change points. Each row include: \code{loc.block} is the change point block index , \code{max.proj.cusum} is the maximum projected cusum value at this change point, \code{depth} is the depth in the wild binary segmentation, \code{s.int} and \code{e.int} are the start and end of the interval on which this change point is detected, \code{loc} is the change point time index \code{loc.block * L}, \code{loc.refine} is the refined change point.}
\item{cpfs}{A list with \code{k} sub-lists, one for each change point, in the order of \code{cp}. Each sub-list include two elements: \code{cpfreq} is the change point frequency at this change point, and \code{pro.allfreq} is a matrix with \code{length(freq)} (or \code{length(band)} subject \code{band} argument setting) rows and \code{m} columns representing the projection at this change point.}
\item{sparsity}{The specified or data-driven sparsity.}
\item{mean}{The piecewise mean of the spectral estimate (or coherency subject to \code{coherency} argument setting), a 4-dimensional order of \code{m} by \code{m} by \code{length(freq)} by \code{k+1}. Could be scaled subject to argument setting.}
}
}
\description{
Change point estimation for multivariate time series from frequency domain
}
\author{
Xinyu Zhang xinyu-zhang@uiowa.edu
}
