% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{SpecCpOverlap}
\alias{SpecCpOverlap}
\title{Function to get a refined change point location using overlap blocks, after got change point block locations. The parameters include some parameters in function \code{SpecBlockOverlapProject}, as well as}
\usage{
SpecCpOverlap(
  y,
  L,
  R = floor(L^(1/3)),
  freq,
  band = rep(1, length(freq)),
  scale = TRUE,
  normaltrans = FALSE,
  coherency = FALSE,
  res,
  bs,
  tau,
  trim
)
}
\arguments{
\item{res}{an internal result by the function \code{wsbs}}

\item{bs}{the non-overlapped block estimate}

\item{tau}{the threshold}

\item{trim}{the boundary removal}
}
\value{
the refined change point location
}
\description{
Function to get a refined change point location using overlap blocks, after got change point block locations. The parameters include some parameters in function \code{SpecBlockOverlapProject}, as well as
}
\keyword{internal}
