% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{SpecProj}
\alias{SpecProj}
\title{Function for frequency-specific projection}
\usage{
SpecProj(bs.part, sparsity = NULL, initial = "aggregation")
}
\arguments{
\item{bs.part}{the four order spectral tensor, of dim by dim by freq (or band) by block (or at)}

\item{sparsity}{the sparsity value}

\item{initial}{the initialization method, can be either aggregation or unfolding.}
}
\value{
a list including information of the projection
}
\description{
Function for frequency-specific projection
}
\keyword{internal}
