% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{Tau.compu}
\alias{Tau.compu}
\title{Function for compute the list of maximum projected cusum  values based on many frequency bootstrap samples, in order to get the threshold}
\usage{
Tau.compu(
  bs = NULL,
  sparsity,
  Nb = 100,
  do.parallel = 0,
  quan = 0.975,
  initial = "aggregation"
)
}
\arguments{
\item{bs}{the block spectral estimate by function \code{SpecBlock}, of dim by dim by freq (or band) by block}

\item{sparsity}{the sparsity value}

\item{Nb}{the number of bootstraps for obtaining the threshold value}

\item{do.parallel}{number of copies of R running in parallel}

\item{quan}{take the \eqn{quan}-th quantile among bootstraps for the threshold value}

\item{initial}{the initialization method, can be either aggregation or unfolding.}
}
\value{
a list containing the threshold in all bootstraps, and the final threshold
}
\description{
Function for compute the list of maximum projected cusum  values based on many frequency bootstrap samples, in order to get the threshold
}
\keyword{internal}
