% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{TensorCusum}
\alias{TensorCusum}
\title{Function for the spectral tensor CUSUM, designed specifically for the symmetric spectral matrix.}
\usage{
TensorCusum(x)
}
\arguments{
\item{x}{4-order tensor of dim by dim by freq (or band) by block (at)}
}
\value{
4-order CUSUM tensor of dim by dim by freq (or band) by block (at)-1
}
\description{
Function for the spectral tensor CUSUM, designed specifically for the symmetric spectral matrix.
}
\keyword{internal}
