% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speccp.R
\name{wsbs}
\alias{wsbs}
\title{Function of wild sparsified binary segmentation}
\usage{
wsbs(
  bs,
  sparsity,
  tau,
  trim = NULL,
  M = NULL,
  do.parallel = 0,
  single = FALSE,
  initial = " aggregation"
)
}
\arguments{
\item{bs}{the four order spectral tensor, of dim by dim by freq (or band) by block}

\item{sparsity}{the sparsity value}

\item{tau}{the threshold}

\item{trim}{the boundary removal \eqn{\nu}}

\item{M}{the number of random intervals in the wild binary segmentation}

\item{do.parallel}{number of copies of R running in parallel}

\item{single}{logical, if TRUE, only a single change point will be estimated, and binary segmentation will not be implemented.}

\item{initial}{the initialization method, can be either aggregation or unfolding.}
}
\value{
a list including the change point location information, and the change point frequency and series information
}
\description{
Function of wild sparsified binary segmentation
}
\keyword{internal}
